Revert "Move ENTITY_CATEGORIES into homeassistant.helpers.entity"
This reverts commit 94742ee048798eb108f58fc2ddbb77c39fd755cf.
